Exception retrieving zip archive length (#2686)
* Resolve exception retrieving length from zip archive stream during scan target enumeration.

* Add renamed test files.

* Reversion to 4.2.1

* Review feedback.

* Elevate code warning.

* Small improvements to unit tests.

* Add comments on too brief test file name in optional data visitor tests.

* Update src/Sarif/GitHelper.cs

Yes! Calls out the mile wide unit test hole. :)


* Add exception filter to unit test.cd..

* Update tests with improved catch. Add git blame test.

* Make git blame test windows only.

* Fix version link in release notes.

* Add 'rel.sarif' explanation to additional test location.

* Improve release note by citing .net 4.8 as a condition for exception.

* Retrieve full file path for git blame.

* Provide test specifically for file location in git helper tests.

* Restore test to working order.

* Make directory location resilient between .net 4.8 and net60

* Remove .net 48 test suppor for now.

---------

+++ src/Test.UnitTests.Sarif.Driver/Sdk/AnalyzeCommandBaseTests.cs
@@ -1025,7 +1025,7 @@ namespace Microsoft.CodeAnalysis.Sarif.Driver
                                                                     It.IsAny<SearchOption>())).Returns(files);
                 mockFileSystem.Setup(x => x.FileOpenRead(It.IsAny<string>())).Returns(mockStream.Object);
                 mockFileSystem.Setup(x => x.FileExists(tempFile.Name)).Returns(true);
-                mockFileSystem.Setup(x => x.FileInfoLength(It.IsAny<string>())).Returns(testCase.fileSizeInBytes);
+                mockFileSystem.Setup(x => x.FileInfoLength(It.IsAny<string>())).Returns((long)testCase.fileSizeInBytes);
 
                 bool expectedToBeWithinLimits =
                     testCase.fileSizeInBytes != 0 &&
